Rework of configuration dump
UTF-8 characters were causing issues on certain systems (daemon could
not start and crashed). This patch completely reworks the formatting of
the configuration dump given by `--dry-run`. This originates in in #11.

Care was taken to ensure that the config dump gives something expectable
in the comments. There was far too things redundant with the actual YAML
configuration in the original format.

Signed-off-by: Matouš Jan Fialka <mjf@mjf.cz>