dockerfile: Do a clean build in the dockerfile
When building a docker container, perform a clean build of the
component within the container instead of relying on the
go toolchain on the invoking machine and whatever the latest
(possibly stale) binary in the directory tree is.

Build dependency caching is enabled to reduce download times.

OCI container labels are added to provide standard container
metadata.

File locations, default invocation etc is the same as the
original Dockerfile.